# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Seu Zuca — Marketplace B2B de Materiais de Construção
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reviews:
    post:
      operationId: createReview
      tags: [reviews]
      summary: Criar avaliação
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewBody"
      responses:
        "201":
          description: Avaliação criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # WISHLIST
...
  /wishlist/{productId}:
    post:
      operationId: addToWishlist
      tags: [wishlist]
      summary: Adicionar produto aos favoritos
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Adicionado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
    delete:
      operationId: removeFromWishlist
      tags: [wishlist]
      summary: Remover dos favoritos
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ADDRESSES
...
  /admin/reviews/{id}/approve:
    post:
      operationId: adminApproveReview
      tags: [admin]
      summary: Aprovar avaliação
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Aprovada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /admin/reviews/{id}/remove:
    delete:
      operationId: adminRemoveReview
      tags: [admin]
      summary: Remover avaliação
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # AUTH SCHEMAS
...
    Review:
      type: object
      properties:
        id:
          type: integer
        buyerId:
          type: integer
        buyerName:
          type: string
        productId:
          type: integer
        productName:
          type: string
        supplierId:
          type: integer
        nota:
          type: integer
        comentario:
          type: string
        aprovada:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - buyerId
        - nota
        - comentario
        - aprovada
        - createdAt
    CreateReviewBody:
      type: object
      properties:
        productId:
          type: integer
        orderId:
          type: integer
        nota:
          type: integer
        comentario:
          type: string
      required:
        - productId
        - orderId
        - nota
        - comentario

    # WISHLIST
...